separate segmentitem and segmentpillbutton to different files
